% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_fetch.R
\name{animal_fetch}
\alias{animal_fetch}
\title{Fetch an animal}
\usage{
animal_fetch(by)
}
\arguments{
\item{by}{(character) which animal to get}
}
\value{
single string
}
\description{
Fetch an animal
}
\examples{
animal_fetch("longtailcat")
cat(animal_fetch("longtailcat"))
animal_fetch("poop")
cat(animal_fetch("poop"))
}
