% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{bubble}
\alias{bubble}
\title{Thought/speech bubble/balloon}
\usage{
bubble(x, width = 60, thought_sym = "o")
}
\arguments{
\item{x}{(character) a character vector}

\item{width}{(integer/numeric) width of each line. default: 60}

\item{thought_sym}{(character) scalar character to use for the
speech bubble tail (see \url{https://en.wikipedia.org/wiki/Speech_balloon}).
default: "o"}
}
\value{
character vector of length greater than the input \code{x}
}
\description{
Thought/speech bubble/balloon
}
\note{
modified from \url{https://github.com/schochastics/startifyR}
}
\examples{
\dontshow{if (rlang::is_installed("fortunes")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(fortunes)
quote <- as.character(fortune())
bubble(x = quote)

text_color <- sample(grDevices::colors(), 1)
text_style <- crayon::make_style(text_color)
text_style(bubble(quote))

cat(bubble(paste(quote, collapse = " ")), sep = "\n")
\dontshow{\}) # examplesIf}
}
