% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{construct.balloon}
\alias{construct.balloon}
\title{Puts the message into the balloon}
\usage{
construct.balloon(message, think)
}
\arguments{
\item{message}{{character vector} a character vector with the message to
put into the bubble (one element per line).}

\item{think}{{logical} is the cow thinking rather than saying?}
}
\value{
{string} a single string with the speech bubble in it (embedded newlines)
}
\description{
Puts the message into the balloon
}
\details{
We expect the message to be already wrapped to the appropriate number of
 lines.

This places the message inside a speech/thought bubble.

If it's a thought bubble, the sides of the bubble are `(` and `)`.

If it's a speech bubble and one-line, the bubble boundaries are `<` and `>`.
If it's a speech bubble and more than one line, the top row has borders
`/` and `\`, the bottom `\` and `/`, and the middle rows `|` and `|`.

(See the examples).
}
\examples{
cat(cowsay:::construct.balloon('MOOOO', think=TRUE))
cat(cowsay:::construct.balloon('MOOOO', think=FALSE))
cat(cowsay:::construct.balloon(c('MOOOO', 'MOOO!!!'), think=FALSE))
}

