% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\docType{data}
\name{cow.styles}
\alias{cow.styles}
\title{Cow styles (borg, dead, etc)}
\format{\preformatted{List of 9
 $ borg    :List of 1
  ..$ eyes: chr "=="
 $ dead    :List of 2
  ..$ eyes  : chr "XX"
  ..$ tongue: chr "U"
 $ default :List of 1
  ..$ eyes: chr "OO"
 $ greedy  :List of 1
  ..$ eyes: chr "$$"
 $ paranoid:List of 1
  ..$ eyes: chr "@@"
 $ stoned  :List of 2
  ..$ eyes  : chr "**"
  ..$ tongue: chr "U"
 $ tired   :List of 1
  ..$ eyes: chr "--"
 $ wired   :List of 1
  ..$ eyes: chr "OO"
 $ young   :List of 1
  ..$ eyes: chr ".."
}}
\usage{
cow.styles
}
\description{
The available styles are:
}
\details{
\describe{
  \item{borg}{the cow looks like a Borg}
  \item{dead}{the cow is dead (!)}
  \item{default}{the default plain cow}
  \item{greedy}{the cow is greedy}
  \item{paranoid}{the cow is paranoid}
  \item{stoned}{the cow is stoned}
  \item{tired}{the cow is sleepy}
  \item{wired}{the cow has had too much caffeine}
  \item{young}{the cow in its younger days}
}

The only differences here are in the eye and tongue configurations.
}
\seealso{
Other cow styles: \code{\link{get.cowfile}};
  \code{\link{get.cowpaths}}; \code{\link{list.cows}}
}
\keyword{datasets}

