% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay-package.r
\docType{package}
\name{cowsay-package}
\alias{cowsay-package}
\title{An R port of cowsay}
\description{
This is an R version of the original cowsay program written in Perl by Tony Monroe.
}
\examples{
cowsay('Hello R!')
library(fortunes)
cowsay(fortune(), cow='tux', style='wired')
}
\author{
Amy Chan <mathematical.coffee@gmail.com>
}
\references{
\href{http://web.archive.org/20120225123719/http://www.nog.net/~tony/warez/cowsay.shtml}{The official website at the wayback machine} (Archived February 25, 2012)
}
\seealso{
\code{\link{cowsay}}
}
\keyword{package}

