% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{.tail}
\alias{.tail}
\title{Make the tail part of a thought bubble}
\usage{
.tail(max_char_length, thought_sym = "o")
}
\arguments{
\item{max_char_length}{(numeric) length of the maximum line. this is used
to determine how much whitespace padding to add to the left of
\code{thought_sym}}

\item{thought_sym}{(character) scalar character to use for the
speech bubble tail (see \url{https://en.wikipedia.org/wiki/Speech_balloon}).
default: "o"}
}
\description{
Make the tail part of a thought bubble
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
.tail(59)
cat(.tail(59), sep = "\n")
cat(.tail(11), sep = "\n")
cat(.tail(11, "\%"), sep = "\n")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
