% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endless_horse.R
\name{endless_horse}
\alias{endless_horse}
\title{Endless horse}
\usage{
endless_horse(what = "Hello world!", endless = TRUE, n = 10)
}
\arguments{
\item{what}{(character) What do you want to say? See details.}

\item{endless}{(logical) Should horse be enless, you better say yes. Default: TRUE}

\item{n}{(integer) Number of leg segments. Default: 10}
}
\description{
Endless horse
}
\examples{
\dontrun{
endless_horse()
endless_horse(endless = FALSE)
endless_horse(n = 15)
}
}

