% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{get.cow}
\alias{get.cow}
\title{Reads a cow from a cowfile, substituting the eyes, thoughts and tongue.}
\usage{
get.cow(cowfile, eyes, thoughts, tongue)
}
\arguments{
\item{cowfile}{{string} path to the cowfile}

\item{eyes}{{string} characters to use for the eyes
(only the first two characters are used), e.g. "oo"}

\item{thoughts}{{string} used in drawing the speech bubble, e.g. "o" or "\\"}

\item{tongue}{{string} characters to use for the tongue
(at most the first two characters are used), e.g. "U", "\/"}
}
\value{
a character string with the cow (with embedded newlines), and
  eyes, thoughts, tongue substituted in.
}
\description{
Returns an error if the cow does not exist.
}
\seealso{
Other cowfile parsing: \code{\link{read.cow.r}}
}

