% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{get.cowfile}
\alias{get.cowfile}
\title{Looks for a particular cow.}
\usage{
get.cowfile(cow)
}
\arguments{
\item{cow}{{string} path to a cow file to use, or name of a pre-installed cow (see
\code{\link{list.cows}})}
}
\value{
{character} the path to the first matching cow on the cowpath, throwing an error if not found.
}
\description{
Looks for cows in this order:
}
\details{
1. If the input is a filename with extension '.rcow' and exists, we use it.
2. Otherwise find the corresponding `{cow}.rcow` on the cowpath.

Note that the `cow` should NOT have the extension
 (i.e. "three-eyes", not "three-eyes.cow")
}
\examples{
get.cowfile(system.file('cows/default.rcow', package='cowsay'))
get.cowfile('three-eyes')
\dontrun{
get.cowfile('nonexistentfoobarbaz')
}
}
\seealso{
\code{\link{get.cowpaths}} to see the cowpath.

Other cow styles: \code{\link{cow.styles}};
  \code{\link{get.cowpaths}}; \code{\link{list.cows}}
}

