% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{get.cowpaths}
\alias{get.cowpaths}
\title{Returns the search path for cows.}
\usage{
get.cowpaths()
}
\value{
{character vector} directories that cows will be looked for under.
}
\description{
* if environment variable `$COWPATH` is set, we use that.
* we always include the cows installed with this package.
}
\details{
Order matters; if a particular cow is found in one of the earlier paths, we
 will not bother looking at the later paths.
}
\seealso{
Other cow styles: \code{\link{cow.styles}};
  \code{\link{get.cowfile}}; \code{\link{list.cows}}
}

