% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{get.message}
\alias{get.message}
\title{Attempts to convert `x` to character better than `as.character`.}
\usage{
get.message(x, store.print.method = FALSE, ...)
}
\arguments{
\item{x}{{character vector} thing to convert to character.}

\item{store.print.method}{{boolean} whether to store the name of the 'print'
method used, if any (default `FALSE`).}

\item{...}{passed on to the `print` method (e.g. a `width` argument)}
}
\value{
{character vector}
 If `x` is a character, returns `x`.
 If `x` has a 'print' method ([`has.print`][has.print]), we capture the output
  of `print(x)`. This may have multiple elements (one per line of output).
 Otherwise, we use `as.character`.
If `store.print.method=TRUE`, the return value has an attribute
'print.method' with the name of the method (see `::has.print`).
}
\description{
Attempts to convert `x` to character better than `as.character`.
}
\details{
The idea is to get a nicely-printed cowsay message regardless of input class.
}
\examples{
m <- lm(Sepal.Length ~ Species, iris)
# compare
cat(as.character(m), sep='\\n') # with
cat(cowsay:::get.message(m), sep='\\n')
}
\seealso{
Other utility functions: \code{\link{has.print}};
  \code{\link{wrap.message}}
}

