% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{has.print}
\alias{has.print}
\title{Does this object have a 'print' method?}
\usage{
has.print(x, return.method = FALSE)
}
\arguments{
\item{x}{{anything} an object that we will query for a print method.}

\item{return.name}{{boolean} whether to return the method name, if one is found.}
}
\value{
{boolean|character} If `x` has no `print` method, returns `FALSE`.
  If `x` *does* have a `print` method, returns either `TRUE` (if `return.name=FALSE`,
  the default) or the name of the method (`return.name=TRUE`).
}
\description{
Does this object have a 'print' method?
}
\details{
This checks for the existence of an S3 or S4 `print` method on any of the
 classes of `x`.
}
\examples{
# this returns TRUE as print.lm exists
has.print(lm(Sepal.Length ~ Species, iris))
# this returns TRUE because although print.mlm doesn't exist, print.lm *does*.
has.print(lm(cbind(Petal.Length, Petal.Width) ~ Species, iris))
}
\seealso{
[`methods`](utils::methods)

Other utility functions: \code{\link{get.message}};
  \code{\link{gsubv}}; \code{\link{wrap.message}}
}

