% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{is.perl.cow}
\alias{is.perl.cow}
\title{Is a cow a Perl-cow?
A cow is a perl cow if `$the_cow` can be found in the file (rudimentary check!) and the extension is '.cow'.}
\usage{
is.perl.cow(cowfile)
}
\arguments{
\item{cowfile}{{string} path to the cowfile}
}
\value{
{boolean} whether the cow is a Perl cow or not.
}
\description{
Is a cow a Perl-cow?
A cow is a perl cow if `$the_cow` can be found in the file (rudimentary check!) and the extension is '.cow'.
}
\seealso{
Other cowfile parsing: \code{\link{get.cow}};
  \code{\link{read.cow.noperl}};
  \code{\link{read.cow.perl}};
  \code{\link{read.cow.plain}}; \code{\link{read.cow.r}}
}

