% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_color.R
\name{multi_color}
\alias{multi_color}
\title{Multi-color text}
\usage{
multi_color(txt = NULL, colors = c("red", "orange", "yellow", "green",
  "blue", "purple"), as_string = FALSE, ...)
}
\arguments{
\item{txt}{(character) Some text to color.}

\item{colors}{(character) A vector of colors, defaulting to 
c("red", "orange", "yellow", "green", "blue", "purple")}

\item{as_string}{(logical) If TRUE, the response is a plain, 
uncolored string. If FALSE, the string is passed to \code{cat}
and colored text is returned.}

\item{...}{Further args.}
}
\value{
A string if \code{as_string} is TRUE, or colored 
text if FALSE.
}
\description{
Multi-color text
}
\details{
This function evenly (ish) divides up your string into 
these colors in the order they appear in \code{colors}.

Any colors in \code{colors()} or hex values (see \code{?rgb})
are fair game.
}
\examples{
multi_color("ahoy")

animals[["buffalo"]] \%>\% 
  multi_color("green", "white", "orange")

rms \%>\% 
  multi_color(sample(colors(), 10))
}
