% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{randomcowsay}
\alias{randomcowsay}
\title{cowsay with a random cow.}
\usage{
randomcowsay(message, cow = NULL, style = NULL, think = NULL,
  rude = FALSE, ...)
}
\arguments{
\item{message}{{string} what to say}

\item{cow}{{string} path to a cow file to use, or name of a pre-installed cow (see
\code{\link{list.cows}})}

\item{style}{{string} a predefined style for the cow (e.g. 'dead' sets the cow's eyes
              to 'XX'). Overrides any eye string passed in.
              By default, no style is applied.
              See details for more styles.}

\item{think}{{logical} is the cow thinking rather than saying?}

\item{rude}{{boolean} whether to include "rude" cows or not.}

\item{...}{passed to \code{\link{cowsay}}.}
}
\description{
The cow, style, and think-style are all randomized (the eyes/tongue are not,
 but if the style gets randomized then that often specifies the eyes/tongue).
}
\details{
If any of the arguments `cow`, `style` or `think` are given, this overrides
 the randomness for that argument.

Note that if you supply `eyes` or `tongue`, the style overrides these if it
 has prespecified values for them.
}
\examples{
randomcowsay('MOOOOO')
}
\seealso{
Other cowsay: \code{\link{cowsay}}
}

