% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{read.cow.perl}
\alias{read.cow.perl}
\title{Read a cowfile (original Perl format)}
\usage{
read.cow.perl(cowfile, eyes, thoughts, tongue, perl = Sys.which("perl")[1])
}
\arguments{
\item{cowfile}{{string} path to the cowfile}

\item{eyes}{{string} characters to use for the eyes
(only the first two characters are used), e.g. "oo"}

\item{thoughts}{{string} used in drawing the speech bubble, e.g. "o" or "\\"}

\item{tongue}{{string} characters to use for the tongue
(at most the first two characters are used), e.g. "U", "\/"}

\item{perl}{{string}: path to the Perl executable (by default
`Sys.which('perl')[1]`)}
}
\value{
a character string with the cow (with embedded newlines), and
  eyes, thoughts, tongue substituted in.
}
\description{
These are cow files from the original cowsay distribution, written in Perl.
We run them through a perl interpreter (if you have one installed); if you
don't have Perl installed we throw an error.
}
\examples{
\dontrun{
# if you have the original cowsay installed on your system...
cowfile <- '/usr/share/cowsay/cows/three-eyes.cow'
cat(cowsay:::read.cow.perl(cowfile, eyes="Oo", thoughts="\\\\", tongue="U"))
#         \\  ^___^
#          \\ (Ooo)\\_______
#            (___)\\       )\\/\\
#             U  ||----w |
#                 ||     ||
}
}
\seealso{
Other cowfile parsing: \code{\link{get.cow}};
  \code{\link{is.perl.cow}}; \code{\link{read.cow.noperl}};
  \code{\link{read.cow.plain}}; \code{\link{read.cow.r}}
}

