% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{replace.tabs}
\alias{replace.tabs}
\title{Replace tabs with the appropriate number of spaces (one tab is 8 spaces)}
\usage{
replace.tabs(x, indent = 0)
}
\arguments{
\item{x}{{character vector} character vector to replace tabs. Each element is
treated as a separate line; embedded newlines are not.}

\item{indent}{{integer} if the string will be indented by some number of characters,
indicate so here. This will affect how the tabs will be calculated!
The indent is not actually added on to the input string, though.}
}
\value{
{character vector} a character vector of the same length as `x` with all
       '\t' replaced by the number of spaces they were to take up if you `cat`-ed
       `x`

If you split up your terminal into columns of width 8, a '\t' puts the next
 bit of content at the start of the next column. So the '\t' consumes the
 number of spaces needed until the string before the '\t' is of a length divisible
 by 8. If the string is already a length divisible by 8, we add 8 spaces.

It should be the case that `cat(x)` and `cat(replace.tabs(x))` appear the
 same width on the terminal (though the former will have less characters than
 the latter if it has a tab that is replaced by spaces in the latter).
}
\description{
Replace tabs with the appropriate number of spaces (one tab is 8 spaces)
}
\examples{
replace.tabs('1234567\\t9') # "1234567 9"
# the indent of 2 would push the 1234567 into the next 'column', so the 9
#  is pushed out to the next column again (character 17)
replace.tabs('1234567\\t9', indent=2) # "1234567         9"
}
\seealso{
Other utility string functions: \code{\link{gsubv}};
  \code{\link{strpad}};
  \code{\link{strwrap.preserve.space}}
}

