% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say_think}
\alias{say_think}
\title{say/think factory}
\usage{
say_think(thought_sym, say_or_think)
}
\arguments{
\item{thought_sym}{(character) scalar character to use for the
speech bubble tail (see \url{https://en.wikipedia.org/wiki/Speech_balloon}).
default: "o"}

\item{say_or_think}{(character)}
}
\description{
say/think factory
}
\keyword{internal}
