% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{strpad}
\alias{strpad}
\title{Pads a string out to a particular width (with spaces).}
\usage{
strpad(x, width, left = TRUE)
}
\arguments{
\item{x}{{character vector} string(s) to pad}

\item{width}{{integer} width to pad out to, of same length as `x` or recycled out to this length.}

\item{left}{{boolean} whether to left-justify (the default) or right-justify.}
}
\value{
{character vector} A vector of same length as `x` where `nchar(output) == width`.
}
\description{
Pads a string out to a particular width (with spaces).
}
\details{
Just a wrapper round `sprintf`. Note we can't use `format(x, width=width)` because
 this counts escape characters like `\t` as two characters ('\' and 't'), rather
 than 1, and hence does not add enough padding.

It should always be that `all(nchar(strpad(x, width)) == width)`.
}
\examples{
x <- '123\\t456\\t789' # nchar(x) == 11
strpad(x, width=15)
nchar(strpad(x, width=15)) == 15
# compare to:
format(x, width=15)
nchar(format(x, width=15)) # 13!
}
\seealso{
[`format`](base::format), [`sprintf`](base::sprintf)

Other utility string functions: \code{\link{gsubv}};
  \code{\link{replace.tabs}};
  \code{\link{strwrap.preserve.space}}
}

