% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{strwrap.preserve.space}
\alias{strwrap.preserve.space}
\title{Wraps a string to a particular width, preserving spaces (except some trailing ones)}
\usage{
strwrap.preserve.space(x, width)
}
\arguments{
\item{x}{{character vector} string(s) to wrap}

\item{width}{{integer} maximum width of each line}
}
\value{
{character vector} a character vector with one element per line of `x`,
   after wrapping.
}
\description{
Wraps a string to a particular width, preserving spaces (except some trailing ones)
}
\details{
Like R's strwrap but does not destroy whitespace as that one does.
In fact, it's a good deal dumber than strwrap.

It's a bit quirky, too: it will drop trailing whitespace, but **only** if keeping
 it would have made the line longer than `width`.

I'm not really convinced on it.

The idea is that you can `cowsay` things where spaces should stay the same, like
 tables. Though, I suppose you should have `wrap=F` for these.

Note: tabs are not substituted with the appropriate number of spaces, if you want
 that you need to call [replace.tabs][] first.
}
\examples{
# @TODO
}
\seealso{
[`strwrap`](base::strwrap)

Other utility string functions: \code{\link{gsubv}};
  \code{\link{replace.tabs}}; \code{\link{strpad}}
}

