% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{trim.message}
\alias{trim.message}
\title{Trims a message to a particular width.}
\usage{
trim.message(x, width = 0.8 * getOption("width"))
}
\arguments{
\item{x}{a character vector, or an object which can be converted to a
    character vector by \code{\link{as.character}}.}

\item{width}{a positive integer giving the target column for wrapping
    lines in the output.}
}
\value{
{character vector} a vector of `x` chunked up into lines of approximately
  width `width`.
  If there were embedded newlines in `x`, the input is also split up according
  to this.
}
\description{
Trims a message to a particular width.
}
\examples{
cowsay:::trim.message('I do not like green eggs and ham.\\nI do not like them, Sam I Am!',
             width=10)
}
\seealso{
\code{\link[base]{strwrap}}

Other speech bubble functions: \code{\link{construct.balloon}}
}

