% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\name{animals}
\alias{animals}
\title{Animals}
\description{
Named vector of animals
}
\details{
\code{animals} is a named character vector of animals,
with each element a character string of variable length specifying
an ASCII animal. Note that some have unicode characters that won't
play well on some operating systems.
}
\examples{
animals
cat(animal_fetch('cow'))
cat(animal_fetch('chicken'))
cat(animal_fetch('chuck'))
cat(animal_fetch('clippy'))
cat(animal_fetch('poop'))
cat(animal_fetch('bigcat'))
}
