% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\docType{data}
\name{animals}
\alias{animals}
\title{Animals}
\format{
An object of class \code{character} of length 56.
}
\usage{
animals
}
\description{
Named vector of animals
}
\details{
\code{animals} is a named character vector of animals,
with each element a character string of variable length specifying
an ASCII animal. Note that some have unicode characters that won't
play well on some operating systems.
}
\examples{
cat(animals['cow'])
cat(animals['chicken'])
cat(animals[['chuck']])
cat(animals[['clippy']])
cat(animals[['poop']])
cat(animals[['bigcat']])
for (animal in animals) cat(animal, sep = "\n")
}
\keyword{datasets}
