% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.r
\docType{data}
\name{animals}
\alias{animals}
\title{Animals}
\format{
An object of class \code{character} of length 49.
}
\usage{
animals
}
\description{
Named vector of animals
}
\details{
\code{animals} is a named character vector of animals,
with each element a character string of variable length specifying
an ASCII animal. Note that some have unicode characters that won't
play well on some operating systems.
}
\examples{
animals['cow']
animals['clippy']
animals[['clippy']]
}
\keyword{datasets}
