% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowsay-package.R
\docType{package}
\name{cowsay-package}
\alias{cowsay}
\alias{cowsay-package}
\title{cowsay: Messages, Warnings, Strings with Ascii Animals}
\description{
Allows printing of character strings as messages/warnings/etc. with ASCII animals, including cats, cows, frogs, chickens, ghosts, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sckott/cowsay}
  \item \url{https://sckott.github.io/cowsay/}
  \item Report bugs at \url{https://github.com/sckott/cowsay/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{myrmecocystus@gmail.com}

Authors:
\itemize{
  \item Amanda Dobbyn \email{amanda.e.dobbyn@gmail.com}
}

Other contributors:
\itemize{
  \item Tyler Rinker [contributor]
  \item Thomas Leeper [contributor]
  \item Noam Ross [contributor]
  \item Rich FitzJohn [contributor]
  \item Carson Sievert [contributor]
  \item Kiyoko Gotanda [contributor]
  \item Andy Teucher [contributor]
  \item Karl Broman [contributor]
  \item Franz-Sebastian Krah [contributor]
  \item Lucy D'Agostino McGowan [contributor]
  \item Guangchuang Yu [contributor]
  \item Philipp Boersch-Supan [contributor]
  \item Andreas Brandmaier [contributor]
  \item Marion Louveaux [contributor]
  \item David Schoch [contributor]
}

}
\keyword{internal}
