% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{cowsay}
\alias{cowsay}
\title{calls cowsay}
\usage{
cowsay(message, cow = "default", eyes = "oo", tongue = "  ",
  wrap = getOption("width"), think = F, style = c("borg", "dead",
  "default", "greedy", "paranoid", "stoned", "tired", "wired", "young"))
}
\arguments{
\item{message}{{string} what to say. If it's encoded, ensure you set the encoding!}

\item{cow}{{string} path to a cow file to use, or name of a pre-installed cow (see
\code{\link{list.cows}})}

\item{eyes}{{string} characters to use for the eyes
(only the first two characters are used), e.g. "oo"}

\item{tongue}{{string} characters to use for the tongue
(at most the first two characters are used), e.g. "U", "\/"}

\item{wrap}{{integer} number of characters to wrap the message at (-1 not to wrap)
           R's `strwrap` (which we use) destroys whitespace, so if you
            are saying ASCII-art or aligned text, use `wrap=F`.

           If not explicity specified, this will set `wrap=-1`
            if the input `message` has a specific `print` method (so that, for
            example, `cowsay`-ing the results of a t-test will not wrap and
            mess up the output).}

\item{think}{{logical} is the cow thinking rather than saying?}

\item{style}{{string} a predefined style for the cow (e.g. 'dead' sets the cow's eyes
              to 'XX'). Overrides any eye string passed in.
              By default, no style is applied.
              See details for more styles.}
}
\value{
the cow-string, invisibly.
}
\description{
calls cowsay
}
\details{
The available cow styles are:

\describe{
  \item{borg}{the cow looks like a Borg}
  \item{dead}{the cow is dead (!)}
  \item{default}{the default plain cow}
  \item{greedy}{the cow is greedy}
  \item{paranoid}{the cow is paranoid}
  \item{stoned}{the cow is stoned}
  \item{tired}{the cow is sleepy}
  \item{wired}{the cow has had too much caffeine}
  \item{young}{the cow in its younger days}
}
}
\examples{
cowsay('mooooo')
cowsay('moooo', eyes='Oo')

# Example with encoded strings (c-cedilla)
x <- "fa\\xE7ile"
# cowsay(x) # will give a warning about invalid locale
Encoding(x) <- 'latin1'
cowsay(x)

# Don't wrap, or the spacing in the summary is killed
m <- lm(Sepal.Length ~ Species, iris)
cowsay(capture.output(summary(m)), wrap=F)

# let's cowsay some more interesting things.
# Note this doesn't wrap even though the length might exceed your `getOption('wrap')`.
#  This is done automagically to avoid messed up output.
cowsay(head(mtcars))
# if `wrap` is explicity provided, we can override that.
cowsay(head(mtcars), wrap=60)

\dontrun{
# every .Rprofile should have this..
library(fortunes)
cowsay(fortune())
}
}
\seealso{
Other cowsay: \code{\link{randomcowsay}}
}

