% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowsay-package.r
\docType{package}
\name{cowsay}
\alias{cowsay}
\alias{cowsay-package}
\title{Sling messages and warnings with flair}
\description{
If you are familiar with `cowsay` on the cli, then you know what this 
is, but for R.  If not, read below.  Why?  Why not?
}
\details{
Contributors:
\itemize{
 \item Scott Chamberlain
 \item Tyler Rinker
 \item Thomas Leeper
 \item Noam Ross
 \item Rich FitzJohn
 \item Kiyoko Gotanda
 \item Carson Sievert
 \item Andy Teucher
 \item Karl Broman
 \item Franz-Sebastian Krah
 \item Lucy D'Agostino McGowan
 \item Guangchuang Yu
 \item Paolo Sonego
 \item Philipp Boersch-Supan
}

There is only one function in the cowsay package. Simplicity for the win.

Say something:

say('hello world!')
}
