% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowsay-package.r
\docType{package}
\name{cowsay}
\alias{cowsay}
\title{Sling messages and warnings with flair}
\description{
If you are familiar with \code{cowsay} on the cli, then you know what this
is, but for R.  If not, read below.  Why?  Why not?
}
\details{
Contributors:
\itemize{
\item Amanda Dobbyn
\item Andreas Brandmaier
\item Andy Teucher
\item Carson Sievert
\item Franz-Sebastian Krah
\item Guangchuang Yu
\item Karl Broman
\item Kiyoko Gotanda
\item Lucy D'Agostino McGowan
\item Noam Ross
\item Paolo Sonego
\item Philipp Boersch-Supan
\item Rich FitzJohn
\item Scott Chamberlain
\item Thomas Leeper
\item Tyler Rinker
\item Marion Louveaux
}

There is only one function in the cowsay package. Simplicity for the win.

Say something:

say('hello world!')
}
