% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.r
\name{cowsayOptions}
\alias{cowsayOptions}
\title{Querying/setting cowsay options}
\usage{
cowsayOptions(o, value)
}
\arguments{
\item{o}{option name (string). See below.}

\item{value}{value to assign (optional)}
}
\description{
The following `cowsay` options are avilable:
* `rude.cows`: character vector (default: `c('sodomized', 'sodomized-sheep', 'head-in')`)
}
\details{
Note: this code is shamelessly adapted from the `panderOptions` function in the
[pander](http://rapporter.github.io/pander/) package; it's just good and how
I would have done it anyway. Thanks muchly, and kudos to the pander team.
}
\examples{
\dontrun{
cowsayOptions()
cowsayOptions('rude.cows')
cowsayOptions('rude.cows', c(cowsayOptions('rude.cows'), 'kiss'))
}
}
\references{
The corresponding function (`panderOptions`) of the pander package: https://github.com/Rapporter/pander/blob/master/R/options.R
}

