% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endless_horse.R
\name{endless_horse}
\alias{endless_horse}
\title{Endless horse}
\usage{
endless_horse(what = "Hello world!", endless = TRUE, wait = 0.5)
}
\arguments{
\item{what}{(character) What do you want to say? See details.}

\item{endless}{(logical) Should horse be enless, you better say yes. Default: TRUE}

\item{wait}{How long to wait between leg segments (time grows
geometrically after the first iteration in order to keep the horse
on screen for a while, but it \emph{will} keep going forever.  Or
until you hit escape/Ctrl-C depending on your platform).}
}
\description{
Each time you press enter, the horse keeps going...and going...
}
\examples{
\dontrun{
endless_horse()
endless_horse(endless = FALSE)
}
}

