% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{get.cow}
\alias{get.cow}
\title{Reads a cow from a cowfile.}
\usage{
get.cow(cowfile, eyes, thoughts, tongue)
}
\arguments{
\item{cowfile}{{string} path to the cowfile}

\item{eyes}{{string} characters to use for the eyes
(only the first two characters are used), e.g. "oo"}

\item{thoughts}{{string} used in drawing the speech bubble, e.g. "o" or "\\"}

\item{tongue}{{string} characters to use for the tongue
(at most the first two characters are used), e.g. "U", "\/"}
}
\value{
a character string with the cow (with embedded newlines), and
  eyes, thoughts, tongue substituted in.
}
\description{
Reads a cow from a cowfile.
}
\details{
1. If the file has extension `rcow`, we read it in as an R cow (\code{\link{read.cow.r}}).
2. Otherwise, if the file assigns to `$the_cow` at some point, we assume it
     is a Perl cow and read it in with \code{\link{read.cow.perl}} or
     \code{\link{read.cow.noperl}} depending if you have Perl installed or not.
3. Otherwise we just read it in as a plain cow (\code{\link{read.cow.plain}}).
}
\seealso{
Other cowfile parsing: \code{\link{is.perl.cow}};
  \code{\link{read.cow.noperl}};
  \code{\link{read.cow.perl}};
  \code{\link{read.cow.plain}}; \code{\link{read.cow.r}}
}

