% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cowsay.r
\name{list.cows}
\alias{list.cows}
\title{Lists the currently-installed cows.}
\usage{
list.cows(path = NULL, rude = cowsayOptions("rude"), ...)
}
\arguments{
\item{path}{{character vector|NULL} path(s) to list the cows under. If NULL,
the cowpath is used (see \code{\link{get.cowpaths}}).}

\item{rude}{{boolean} whether to include "rude" cows or not.}

\item{...}{passed to \code{\link[base]{list.files}}.}
}
\value{
a character vector of cow files found in the path.
}
\description{
This lists all the `*.cow`, `*.rcow` files found under `path`.
}
\seealso{
\code{\link[base]{list.files}}; \code{\link{cowsayOptions}} to change whether rude cows are included by default, and which cows are rude.

Other cow styles: \code{\link{cow.styles}};
  \code{\link{get.cowfile}}; \code{\link{get.cowpaths}}
}

