% Generated by roxygen2 (4.0.0): do not edit by hand
\name{say}
\alias{say}
\title{Sling messages and warnings with flair}
\usage{
say(what = "Hello world!", by = "cow", type = "message")
}
\arguments{
\item{what}{What do you want to say? See details.}

\item{by}{Type of thing, one of cow, chicken, poop, cat, ant, pumpkin,
ghost, spider, rabbit, pig, snowman, frog, hypnotoad, or longcat. We use match.arg
internally, so you can use unique parts of words that don't conflict with others,
like "g" for "ghost" because there's no other animal that starts with "g".}

\item{type}{One of message (default), warning, or string (returns string)}
}
\description{
Sling messages and warnings with flair
}
\details{
You can put in any phrase you like, OR you can type in one of a few special phrases
that do particular things. They are:

\itemize{
 \item catfact A random cat fact from http://catfacts-api.appspot.com/doc.html
 \item iheart A random quote from http://iheartquotes.com/api
 \item fortune A random quote from an R coder, from fortunes library
 \item time Print the current time
}

Note that if you choose \code{by='hypnotoad'} the quote is forced to be, as you could imagine,
'All Glory to the HYPNO TOAD!'.
}
\examples{
say()
say("what")
say('time')
say('time', "poop")
say("who you callin chicken", "chicken")
say("ain't that some shit", "poop")
say("icanhazpdf?", "cat")
say("boo!", "pumpkin")
say("hot diggity", "frog")
say("fortune", "spider")

# Vary type of output, default calls message()
say("hell no!")
say("hell no!", type="warning")
say("hell no!", type="string")

# Using fortunes
say(what="fortune")

# Using catfacts
say("catfact", "cat")

# Using iheartquotes
say("iheart", "chicken")

# The hypnotoad
say(by="hypnotoad")

# The longcat
say("it's caturday", "longcat")
}

